etl/app/pyproject.toml
[tool.poetry]
name = "sprint 2. django api"
version = "0.1.0"
description = "sprint 2 yandex practicum middle python"
authors = ["Sergey Bilibin <s_bilibin2017@yandex.ru.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django="==3.2.4"
django-split-settings="==1.2.0"
django-postgres-extra="==2.0.6"
djangorestframework="==3.14.0"
gunicorn="==20.1.0"
python-dotenv="==0.21.1"
psycopg2-binary="==2.9.5"
django-extensions="==3.2.1"
django-debug-toolbar="==3.8.1"
requests="==2.28.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
warn_unused_configs = true
disallow_incomplete_defs = true